fix: problem with farmingTimer not removing properly

+++ game/plugins/src/main/kotlin/gg/rsmod/plugins/content/skills/farming/core/PlayerManager.kt
@@ -51,7 +51,7 @@ object PlayerManager {
                     break
                 }
             }
-            player.timers[Constants.playerFarmingTimer] = ticksLeftOnNextTimer.toInt()
+            player.timers[Constants.playerFarmingTimer] = ticksLeftOnNextTimer
 
             // Set the player attribute that stores the last world farm tick the player handled
             // If the current farming tick was not included in the replay, this is the world farming tick minus 1
